+++ main-yt-dlp.py
@@ -25,7 +25,7 @@ except ImportError:
     SLACK_WEBHOOK = None
 
 # Version corresponds to the latest changelog entry timestamp
-VERSION = '2026-01-11-1900'
+VERSION = '2026-02-02-1430'
 
 logger = logging.getLogger(__name__)